New normalizer implementations and fixes
- Add `reinhard_fast_mask` numpy implementation
- Add `vahadane_sklearn` and `vahadane_spams` implementations. For performance (and dependency) reasons, the sklearn implementation of Vahadane is default.
- Add default normalizer fits to `sf.norm.utils.fit_presets`. Normalizers will load these presets instead of fitting to `sf.norm.norm_tile.jpg`. The `v1` preset was generated by fitting each normalizer to `sf.norm.norm_tile.jpg` (thus giving identical results to prior versions). The `v2` preset is a hand-curated optimized normalization which will become default in version 1.3.
- Normalizers can be set to a preset by passing the name of the preset to `.fit()`. Preset names can also be passed as a hyperparameter to `normalizer_source`.
- Merged
'reinhard', `reinhard_fast` and `reinhard_mask` files for simplicity.
- Add Tensorflow-native implementations of `reinhard_mask` and `reinhard_fast_mask`
- Add testing and fixes
