+++ slideflow/norm/tensorflow/__init__.py
@@ -4,7 +4,7 @@ import numpy as np
 from slideflow import errors
 from slideflow.dataset import Dataset
 from slideflow.norm import StainNormalizer
-from slideflow.norm.tensorflow import reinhard, reinhard_fast, macenko
+from slideflow.norm.tensorflow import reinhard, macenko
 from slideflow.util import detuple, log
 from tqdm import tqdm
 
@@ -15,7 +15,9 @@ class TensorflowStainNormalizer(StainNormalizer):
 
     normalizers = {
         'reinhard': reinhard.ReinhardNormalizer,
-        'reinhard_fast': reinhard_fast.ReinhardFastNormalizer,
+        'reinhard_fast': reinhard.ReinhardFastNormalizer,
+        'reinhard_mask': reinhard.ReinhardMaskNormalizer,
+        'reinhard_fast_mask': reinhard.ReinhardFastMaskNormalizer,
         'macenko': macenko.MacenkoNormalizer
     }
 
@@ -123,7 +125,7 @@ class TensorflowStainNormalizer(StainNormalizer):
                 deviations for the normalizer. May raise an error if the
                 normalizer does not have a target_stds fit attribute.
         """
-        self.n.set_fit(**kwargs)
+        self.n.set_fit(**{k:v for k, v in kwargs.items() if v is not None})
 
     def fit(
         self,
@@ -158,7 +160,11 @@ class TensorflowStainNormalizer(StainNormalizer):
                 total=dataset.num_tiles
             )
             for i, slide in dts:
-                fit_vals = self.n.fit(i, reduce=True)
+                if self.vectorized:
+                    fit_vals = self.n.fit(i, reduce=True)
+                else:
+                    _img_fits = zip(*[self.n.fit(_i) for _i in i])
+                    fit_vals = [tf.reduce_mean(tf.stack(v), axis=0) for v in _img_fits]
                 if all_fit_vals == []:
                     all_fit_vals = [[] for _ in range(len(fit_vals))]
                 for v, val in enumerate(fit_vals):
@@ -168,6 +174,10 @@ class TensorflowStainNormalizer(StainNormalizer):
 
         elif isinstance(arg1, np.ndarray):
             self.n.fit(tf.convert_to_tensor(arg1))
+
+        # Fit to a preset
+        elif isinstance(arg1, str) and arg1 in ('v1', 'v2'):
+            self.n.fit_preset(arg1)
 
         elif isinstance(arg1, str):
             self.src_img = tf.image.decode_jpeg(tf.io.read_file(arg1))
@@ -190,7 +200,7 @@ class TensorflowStainNormalizer(StainNormalizer):
         batch: Union[Dict, tf.Tensor],
         *args: Any
     ) -> Union[Dict, tf.Tensor, Tuple[Union[Dict, tf.Tensor], ...]]:
-        """Normalize a batch of tensors.
+        """Normalize a Tensor image or batch of image Tensors.
 
         Args:
             batch (Union[Dict, tf.Tensor]): Dict of tensors with image batches
